requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
streamlit==1.29.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pandas==2.1.4
requests==2.31.0